Debugging
Tried to switch for using one global sortBy and direction set by
table/chart. The data then would be sorted based on the sortBy/direction
change and passed to the chart and table.

In the current code, the table is not re-rendered when the data are
sorted. This solves returning a copy of the data in fileringUtils, but
it adds issue with the useEffect in the lifecycleTable component which
is dependent on the data (always restart the sort indicators etc.) -
this would be probably possible to solve - mark the reset dependent on
something else, like Life Cycle dropdown or View value (installed, all).

The problem is the useState variables are updated is some weird manner
and I actually don't know what exactly is happening there.