[RSPEED-1185] Fix sorting sync (#165)
* Sync from dropdown to table

* Infinite loop - sync table to chart

experimented, if typescript has any processes to avoid it. I doesn't
have.

* fix the infinite loop

* Sorting sync works but with bugs

* Fix bug on rhel lifecycle

* improve comment

* Fix the sync after dropdown change

Sync table with chart wasn't working when Lifecycle dropdown was
changed. Now it's synchronized.

* Add option for sorting control

* Sort is synchronized

Sorting is synchronized, but except name and systems is working a bit
weird.

* Synchronized, but reversed

* Synchronized

The table and chart are synchronized. Still it behaves weird and
probably contains bugs -> needs more work.

* Debugging

Tried to switch for using one global sortBy and direction set by
table/chart. The data then would be sorted based on the sortBy/direction
change and passed to the chart and table.

In the current code, the table is not re-rendered when the data are
sorted. This solves returning a copy of the data in fileringUtils, but
it adds issue with the useEffect in the lifecycleTable component which
is dependent on the data (always restart the sort indicators etc.) -
this would be probably possible to solve - mark the reset dependent on
something else, like Life Cycle dropdown or View value (installed, all).

The problem is the useState variables are updated is some weird manner
and I actually don't know what exactly is happening there.

* Try to fix the issue

Co-authored-by: Preston Watson <prwatson@redhat.com>

* Fix sorting when switching between dropdowns

Co-authored-by: Preston Watson <prwatson@redhat.com>

* fix unit tests

* Add order URL parameter and apply suggestion

---------

Co-authored-by: Preston Watson <prwatson@redhat.com>